pyproject.toml
[project]
name = "docwise"
version = "0.1.0"
description = ""
authors = [
    {name = "pb"}
]
readme = "README.md"
requires-python =  ">=3.11,<3.13" 
dependencies = [
    "streamlit (>=1.46.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "firecrawl (>=2.15.0,<3.0.0)",
    "groq (>=0.30.0,<0.31.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-groq (>=0.3.6,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
